Разбил страницу конструктора на компоненты
