+++ src/frontend/src/App.vue
@@ -1,6 +1,6 @@
 <template>
   <div id="app">
-    <index />
+    <Index />
   </div>
 </template>
