+++ src/frontend/src/layouts/AppLayout.vue
@@ -1,0 +1,30 @@
+<template>
+  <div>
+    <header class="header">
+      <div class="header__logo">
+        <a href="#" class="logo">
+          <img
+            src="@/assets/img/logo.svg"
+            alt="V!U!E! Pizza logo"
+            width="90"
+            height="40"
+          />
+        </a>
+      </div>
+      <div class="header__cart">
+        <a href="#">0 ₽</a>
+      </div>
+      <div class="header__user">
+        <a href="#" class="header__login"><span>Войти</span></a>
+      </div>
+    </header>
+
+    <slot />
+  </div>
+</template>
+
+<script>
+export default {
+  name: "AppLayout",
+};
+</script>
