+++ src/frontend/src/modules/builder/components/BuilderIngredientsSelector.vue
@@ -1,0 +1,84 @@
+<template>
+  <div class="content__ingredients">
+    <div class="sheet">
+      <h2 class="title title--small sheet__title">Выберите ингредиенты</h2>
+
+      <div class="sheet__content ingredients">
+        <div class="ingredients__sauce">
+          <p>Основной соус:</p>
+
+          <label
+            class="radio ingredients__input"
+            v-for="sauce in sauces"
+            :key="sauce.id"
+          >
+            <input type="radio" name="sauce" :value="sauce.value" checked />
+            <span>{{ sauce.name }}</span>
+          </label>
+        </div>
+
+        <div class="ingredients__filling">
+          <p>Начинка:</p>
+
+          <ul class="ingredients__list">
+            <li
+              class="ingredients__item"
+              v-for="item in ingredients"
+              :key="item.id"
+            >
+              <span
+                class="filling"
+                :class="`filling--${ingredientName(item)}`"
+                >{{ item.name }}</span
+              >
+
+              <div class="counter counter--orange ingredients__counter">
+                <button
+                  type="button"
+                  class="counter__button counter__button--minus"
+                  disabled
+                >
+                  <span class="visually-hidden">Меньше</span>
+                </button>
+                <input
+                  type="text"
+                  name="counter"
+                  class="counter__input"
+                  value="0"
+                />
+                <button
+                  type="button"
+                  class="counter__button counter__button--plus"
+                >
+                  <span class="visually-hidden">Больше</span>
+                </button>
+              </div>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "BuilderIngredientsSelector",
+  props: {
+    sauces: {
+      type: Array,
+      required: true,
+    },
+    ingredients: {
+      type: Array,
+      required: true,
+    },
+  },
+  methods: {
+    ingredientName(item) {
+      const match = item.image.match(/(\w+).svg/);
+      return match?.[1] ?? "";
+    },
+  },
+};
+</script>
