+++ src/frontend/src/modules/builder/components/BuilderSizeSelector.vue
@@ -1,0 +1,36 @@
+<template>
+  <div class="content__diameter">
+    <div class="sheet">
+      <h2 class="title title--small sheet__title">Выберите размер</h2>
+
+      <div class="sheet__content diameter">
+        <label
+          class="diameter__input"
+          v-for="size in sizes"
+          :key="size.id"
+          :class="`diameter__input--${size.value}`"
+        >
+          <input
+            type="radio"
+            name="diameter"
+            :value="size.value"
+            class="visually-hidden"
+          />
+          <span>{{ size.name }}</span>
+        </label>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "BuilderSizeSelector",
+  props: {
+    sizes: {
+      type: Array,
+      required: true,
+    },
+  },
+};
+</script>
